+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * *BREAKING* Display custom fields in different accordions depending on the selected section. Refs UIU-3131.
 * *BREAKING* Update the user record view page to display custom fields under the "User information", "Extended information", "Contact information", or default "Custom fields" accordions as configured in settings. Refs UIU-3129.
 * Only show Keycloak user record confirmation when assigning roles to a non-Keycloak record. Refs UIU-3426.
+* Add 'Minor' column to Patrons Pre-Registration List. Refs UIU-3425.
 
 ## [12.1.7] (https://github.com/folio-org/ui-users/tree/v12.1.7) (2025-06-03)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v12.1.6...v12.1.7)